refactor(hypervisor): split darwin vm
